packages/app/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
attrs==25.1.0
baml-py==0.76.2
baml_core_ffi==0.1.7
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.3
ollama==0.1.9
openai==1.63.2
opencv-python==4.11.0.86
opentelemetry-api==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==23.2
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==7.4.4
pytest-asyncio==0.21.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
regex==2024.11.6
requests==2.32.3
simple-websocket==1.1.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.1
typeguard==4.4.2
types-regex==2023.12.25.20240311
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
